fix: fix the issue of axis's update transition animation not work
